Clarify format of public key at rest
The public key format specified in RFD002 mandated standard
Base64-encoding, but the implementation uses URL-safe Base64. We want to
use the URL-safe format to be consistent with the challenge encoding.

This also introduces a dedicated test for the login module.
