+++ docs/rfd/002.md
@@ -24,9 +24,9 @@ See also [google/glome#100](https://github.com/google/glome/issues/100).
 
 # Design ideas
 
-Public keys are stored in base64 encoding and tagged with their protocol
-variant. The configuration file format accepts keys in a format similar to
-[OpenSSH's `authorized_keys` format][1].
+Public keys are stored in URL-safe base64 encoding and tagged with their
+protocol variant version. The configuration file format accepts keys in a
+format similar to [OpenSSH's `authorized_keys` format][1].
 
 [1]: https://man.openbsd.org/sshd.8#AUTHORIZED_KEYS_FILE_FORMAT
 
@@ -37,8 +37,8 @@ The format of a _GLOME public key_ adheres to the ABNF below:
 ```abnf
 public-key = key-type SP key-base64
 key-type = "glome-v1"
-key-base64 = 44base64-char
-base64-char = "=" / "+" / "/" / ALPHA / DIGIT
+key-base64 = 44urlsafe-base64-char
+urlsafe-base64-char = "=" / "-" / "_" / ALPHA / DIGIT
 ```
 
 The key type encodes the GLOME variant this key should be used with. As we
@@ -47,7 +47,7 @@ only have one variant right now, we're only defining one `key-type` here.
 An example public key, like it would be printed by `glome pubkey`:
 
 ```
-glome-v1 lXmlq5jynG6um/w4D4N13TRIE+x7jt0TKVNDMSRS2/I=
+glome-v1 lXmlq5jynG6um_w4D4N13TRIE-x7jt0TKVNDMSRS2_I=
 ```
 
 ## Public Key Interpretation
